+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add ESC + auto focus support for dialogs
 - Update Vue to 3.2.41
 - Add finder module (only for super admin users)
+- Set default value to false for boolean fields
 
 ## 2.3.2 (2022-10-24)